WEBSITE/requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
cassandra-driver==3.24.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.1.5
Flask==1.1.2
Flask-WTF==0.14.3
gensim==3.8.3
geomet==0.2.1.post1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.6.2
MarkupSafe==1.1.1
numpy==1.19.5
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
scipy==1.6.0
six==1.15.0
smart-open==4.1.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.2
Werkzeug==1.0.1
WTForms==2.3.3
